requirements.txt
absl-py==1.3.0
astunparse==1.6.3
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mypy==0.982
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.21.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pathspec==0.10.1
Pillow==9.3.0
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyflakes==2.5.0
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.1.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
